tj addon compat and combustion sound config
