+++ src/main/java/gregtech/api/metatileentity/WorkableTieredMetaTileEntity.java
@@ -20,6 +20,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraft.util.EnumHand;
 import net.minecraft.util.ResourceLocation;
+import net.minecraft.util.SoundEvent;
 import net.minecraft.util.text.TextComponentString;
 import net.minecraft.world.World;
 import net.minecraftforge.fluids.FluidStack;
@@ -68,6 +69,21 @@ public abstract class WorkableTieredMetaTileEntity extends TieredMetaTileEntity 
             }
         };
     }
+
+
+
+    @Override
+    @SideOnly(Side.CLIENT)
+    public SoundEvent getSound() {
+        return workable.recipeMap.getSound();
+    }
+
+    @SideOnly(Side.CLIENT)
+    @Override
+    public boolean shouldPlaySound() {
+        return isValid() && workable.isActive();
+    }
+
 
     @Override
     @SideOnly(Side.CLIENT)
